+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 * **New** - Gravity Forms phone number validation for international numbers
 * **New** - Gravity Forms phone format added to phone fields
 * **New** - Gravity Forms: Mask for phone field
+* **New** - Gravity Forms: Mask for date field
 
 ### 1.3.7: July 9th, 2021
 * **New** - Gravity Forms: autocomplete off for numbers